+++ src/CognitiveServices/CognitiveServices/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated PowerShell to use 2021-10-01 version.
+* Added CommitmentTier and CommitmentPlan cmdlets.
+* Added Deployment cmdlets.
+* Added `New-AzCognitiveServicesObject` cmdlet for generating CommitmentPlan/Deployment objects.
 
 ## Version 1.9.0
 * [Breaking Change] Changed type of PSCognitiveServicesAccount.Identity.Type from IdentityType to ResourceIdentityType.