+++ src/MySql/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.0.0
+* General availability of Az.MySql
+
 ## Version 0.8.0
 * Changed API version to 2021-05-01
 * Fixed bugs for location capabilities